refactoread get_members

+++ client/src/api/file_request.c
@@ -1,6 +1,19 @@
 #include "client.h"
 
-t_dtp *mx_upload_file_request(const char *name, goffset size, char *token, gint room_id) {
+/*
+ * Function: mx_upload_file_request
+ * -------------------------------
+ * Creates upload file request
+ * 
+ * name: name of file
+ * size: size of file
+ * token: user SHA-256 token
+ * room_id: room id
+ * 
+ * returns: upload file request
+ */
+t_dtp *mx_upload_file_request(const char *name, goffset size,
+                              char *token, guint64 room_id) {
     cJSON *json_result = cJSON_CreateObject();
 
     if (!cJSON_AddNumberToObject(json_result, "type", RQ_UPLOAD_FILE))
